% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box_cox.R
\name{box_cox}
\alias{box_cox}
\title{Function for aplying a Box-Cox transformation}
\usage{
box_cox(x, l)
}
\arguments{
\item{x}{Vector to apply the Box-Cox to}

\item{l}{The parameter lambda}
}
\value{
The vector of the transformed \code{x}.
}
\description{
This function aplies a Box-Cox transformation to a positive vector
}
\details{
\itemize{
  \item The Box-Cox transformation is defined for \code{x>0}.
  \item NA values are ommited.
}
}
\examples{
box_cox(tenvec,1)
box_cox(2:8, 4)
box_cox(1:5, -2)

}
