% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c_root.R
\name{c_root}
\alias{c_root}
\title{Check if a root is a real number}
\usage{
c_root(a, b, c)
}
\arguments{
\item{a}{Constant of the cuadratic term.}

\item{b}{Constant of the linear term.}

\item{c}{Constant term.}
}
\value{
\code{1} if \code{b^2-4*a*c>0}, \code{0} if \code{b^2-4*a*c=0}
\code{-1} if \code{b^2-4*a*c<0}.
}
\description{
The function checks if the solution of the root
when solving a second degree plynomial (\code{a*x^2+b*x+c}) is positive,
i.e. checks if \code{b^2-4*a*c>0}.
This is an internal function.
}
