% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cube.R
\name{cube}
\alias{cube}
\title{cube a vector}
\usage{
cube(x, add_plot = FALSE)
}
\arguments{
\item{x}{The vector to be cubed.}

\item{add_plot}{If a plot with the cubed vector should be plotted
(The default is FALSE)}
}
\value{
A vector that is the cube of \code{x}.
}
\description{
This function cubes a vector.
}
\details{
NA values are be omited.
}
\examples{
cube(1:10,add_plot=TRUE)
cube(1:10,add_plot=FALSE)
cube(-5,add_plot=FALSE)
}
