% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynom.R
\name{polynom}
\alias{polynom}
\title{Solution of a second degree polynomial}
\usage{
polynom(a, b, c)
}
\arguments{
\item{a}{Constant of the cuadratic term.}

\item{b}{Constant of the linear term.}

\item{c}{Constant term.}
}
\value{
A vector with the solution(s)
}
\description{
Finds the solution(s) \code{x} in a second degree
polynomial  (\eqn{a*x^2+b*x+c})
}
\details{
The number of solutions depends on the value of
\eqn{b^2-4*a*c}.

\itemize{
     \item If \deqn{b^2-4*a*c=0} the polynomial has one solution.
     \item If \deqn{b^2-4*a*c>0} the polynomial has two solutions.
     \item If \deqn{b^2-4*a*c<0} the polynomial has no real solution.
}
}
\examples{
polynom(3,6,3)
polynom(4,2,1)
polynom(7,8,2)
}
