% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow.R
\name{pow}
\alias{pow}
\title{Generic Power Function}
\usage{
pow(x, a, add_plot)
}
\arguments{
\item{x}{Vector to raise to some power.}

\item{a}{Power to raise \code{x} by.}

\item{add_plot}{Display a plot of \code{x} vs the output. Use logical.}
}
\value{
The vector \code{x}, raised to the power of \code{a}.
}
\description{
This is an intern function.
This function raises a vector to any power.
This funtion also plots the raised powers.
}
