% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Computation of Powers}
\description{
With this package you can
\itemize{
\item Square a vector \code{square(x,...)}
\item Cube a vector \code{cube(x,..)}
\item Obtain the reciprocal of a vector \code{reciprocal(x,...)}
\item Apply a Box-Cox transformation to a vector \code{box_cox(x,l,..)}
\item Find the solution of a second degree polynomial \code{polynom(a,b,c)}
}
}
\note{
This package is only for class assignment purposes
}
\author{
Mar Bazua
}
