% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reciprocal.R
\name{reciprocal}
\alias{reciprocal}
\title{Compute the reciprocal}
\usage{
reciprocal(x, add_plot = FALSE)
}
\arguments{
\item{x}{The vector to take the reciprocal}

\item{add_plot}{If a plot with the reciprocal vector should be plotted
(The default is FALSE)}
}
\value{
A vector that is the reciprocal of \code{x}.
}
\description{
This function computes the reciprocal for each entry of a vector
}
\details{
\itemize{
  \item The values of the vector must be different to zero.
  \item NA values are be omited.}
}
\examples{
reciprocal(tenvec,add_plot=TRUE)
reciprocal(5,add_plot=FALSE)
reciprocal(-5,add_plot=FALSE)
}
