% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/square.R
\name{square}
\alias{square}
\title{Square a vector}
\usage{
square(x, add_plot = FALSE)
}
\arguments{
\item{x}{The vector to be squared.}

\item{add_plot}{If a plot with the squared vector should be plotted.
(The default is FALSE)}
}
\value{
A vector that is the square of \code{x}.
}
\description{
This function squares a vector.
}
\details{
NA values are be omited.
}
\examples{
square(1:10, add_plot=TRUE)
square(2:8, add_plot=FALSE)
square(-5, add_plot=FALSE)
}
