% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powers-package.r
\docType{package}
\name{powers}
\alias{powers}
\alias{powers-package}
\title{Computation of Powers}
\description{
With this package you can apply powers to your vectors by
using functions
}
\details{
You can also find the solution of a second degree polynomial
}
\note{
This package is only for class assignment purposes
}
\author{
Mar Bazua
}
